refactor

+++ src/kz/lof/dataengine/jpa/DAO.java
@@ -119,7 +119,7 @@ public abstract class DAO<T extends IAppEntity, K> implements IDAO<T, K> {
 			EntityTransaction t = em.getTransaction();
 			try {
 				t.begin();
-				entity.setAuthor(user.getId());
+				entity.setAuthor((long) user.getId());
 				entity.setForm(entity.getDefaultFormName());
 				em.persist(entity);
 				t.commit();
